+++ tests/apps/core/utils.py
@@ -1,9 +1,11 @@
 """Test utils for the courses application"""
+import random
+from functools import reduce
 
 from django.test.client import RequestFactory
 
 from cms.middleware.toolbar import ToolbarMiddleware
-from cms.toolbar.items import Menu, ModalItem
+from cms.toolbar.items import ModalItem
 
 
 class CheckToolbarMixin:
@@ -12,10 +14,18 @@ class CheckToolbarMixin:
     CMS toolbar unit tests
     """
 
+    @staticmethod
+    def _get_items(toolbar, name, item_type):
+        """Get items matching a name and type throughout all menus of a toolbar."""
+        return reduce(
+            lambda acc, menu: acc + menu.find_items(item_type, name=name),
+            toolbar.menus.values(),
+            [],
+        )
+
     def check_active(self, toolbar, name, item_type=ModalItem):
         """The item should be present in the toolbar and active."""
-        page_menu_result = toolbar.find_items(Menu, name="Page")[0]
-        results = page_menu_result.item.find_items(item_type, name=name)
+        results = self._get_items(toolbar, name, item_type)
         self.assertEqual(len(results), 1)
         item = results[0].item
         self.assertFalse(item.disabled)
@@ -23,8 +33,7 @@ class CheckToolbarMixin:
 
     def check_disabled(self, toolbar, name, item_type=ModalItem):
         """The item should be present in the toolbar and disabled."""
-        page_menu_result = toolbar.find_items(Menu, name="Page")[0]
-        results = page_menu_result.item.find_items(item_type, name=name)
+        results = self._get_items(toolbar, name, item_type)
         self.assertEqual(len(results), 1)
         item = results[0].item
         self.assertTrue(item.disabled)
@@ -32,8 +41,7 @@ class CheckToolbarMixin:
 
     def check_missing(self, toolbar, name, item_type=ModalItem):
         """The item should not be in the toolbar."""
-        page_menu_result = toolbar.find_items(Menu, name="Page")[0]
-        results = page_menu_result.item.find_items(item_type, name=name)
+        results = self._get_items(toolbar, name, item_type)
         self.assertEqual(results, [])
 
     # pylint: disable=unused-argument
@@ -42,11 +50,17 @@ class CheckToolbarMixin:
         self.assertEqual(toolbar.get_left_items(), [])
 
     @staticmethod
-    def get_toolbar_for_page(page, user, edit, preview):
+    def get_toolbar_for_page(page, user, edit=None, preview=None):
         """
         This method is a helper to build a request to test the toolbar in different states
         for different users
         """
+        if edit is None:
+            edit = random.choice([True, False])
+
+        if preview is None:
+            preview = random.choice([True, False])
+
         url = page.get_absolute_url()
         factory = RequestFactory()